Implement `ExactSizeIterator` for iterator types
[`ExactSizeIterator`] indicates that an iterator knows its exact length.
Iterator types defined in this crate (`Iter`, `Values`, and `IntoIter`)
iterates on backend `Vec<Box<str>>`, which has exact size info.

Note that this trait impl restricts interner's backend.
For example, if the backend is changed to some other data structure
where it is not cheap to know the number of elements, then keeping
iterators `ExactSizeIterator` can be a bit costly.

[`ExactSizeIterator`]: https://doc.rust-lang.org/stable/std/iter/trait.ExactSizeIterator.html